Add files via upload
LTB Enclosure Front contains the front shell of the enclosure as well as the "shelf" for the PCB that goes inside of the front shell.

LTB Enclosure v8 contains the back shell of the LTB enclosure. It has a kickstand, and thus it will require a lot of support extrusion material and takes about 22 hours to print because of this fact.